+++ tensorflow/core/kernels/data/repeat_dataset_op.cc
@@ -14,6 +14,7 @@ limitations under the License.
 ==============================================================================*/
 #include "tensorflow/core/kernels/data/repeat_dataset_op.h"
 
+#include <cstdlib>
 #include <functional>
 #include <memory>
 #include <optional>
@@ -378,7 +379,7 @@ class RepeatDatasetOp::Dataset : public DatasetBase {
       int64_t input_cardinality = dataset()->input_->Cardinality();
       int64_t repeat_count = i_;
       return [parent_index_mapper, input_cardinality,
-              repeat_count](int64_t element_position) -> int64_t {
+              repeat_count](size_t element_position) -> size_t {
         if (element_position >= input_cardinality) {
           // The input element position is out-of-range. The caller is
           // responsible for handle this case (e.g.: returning end_of_sequence).
@@ -389,9 +390,9 @@ class RepeatDatasetOp::Dataset : public DatasetBase {
         // [0, input_range] to [0, input_range * repetitions].
         // Then, reduces the shuffled indices to [0, input_range] by taking the
         // mod. This way, the shuffling happens across repetitions.
-        int64_t repeated_element_position =
+        size_t repeated_element_position =
             repeat_count * input_cardinality + element_position;
-        int64_t shuffled_element_position =
+        size_t shuffled_element_position =
             parent_index_mapper(repeated_element_position);
         return shuffled_element_position % input_cardinality;
       };
